requirements.txt
requests~=2.31.0
anthropic~=0.25.9
python-dotenv~=1.0.1
